write_mask should be type uint32_t

+++ embedded/src/bert/bert.c
@@ -881,7 +881,7 @@ int bert_accelerated_to_physical(int logical,uint32_t *frame_data,uint64_t *logi
   // mask for bits that should be written for this memory
   //  useful for ramb18 case where u64s may overlap between intended memory bits
   //  and the partner memory bits
-  uint64_t *write_mask=(uint32_t *)malloc(sizeof(uint32_t)*(2*u64_per_lookup));
+  uint32_t *write_mask=(uint32_t *)malloc(sizeof(uint32_t)*(2*u64_per_lookup));
   for (int i=0;i<2*u64_per_lookup;i++)
     {
       write_mask[i]=0;
